requirements.txt
absl-py==1.4.0
alembic==1.11.1
astor==0.8.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
blobfile==2.0.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
clearml==1.12.1rc0
click==8.1.6
cloudpickle==2.2.1
cmaes==0.10.0
cmake==3.27.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
contextlib2==21.6.0
cucim==23.6.0
cycler==0.11.0
databricks-cli==0.17.7
docker==6.1.3
einops==0.6.1
entrypoints==0.4
filelock==3.11.0
fire==0.5.0

flatbuffers==23.5.26
fonttools==4.41.1
fsspec==2023.6.0
furl==2.1.3
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.56.2
gunicorn==20.1.0
h5py==3.9.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imagecodecs==2023.3.16
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
itk==5.3.0
itk-core==5.3.0
itk-filtering==5.3.0
itk-io==5.3.0
itk-numerics==5.3.0
itk-registration==5.3.0
itk-segmentation==5.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
json-tricks==3.17.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lazy_loader==0.3
lit==16.0.6
lmdb==1.4.1
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
MedPy==0.4.0
mlflow==2.5.0
monai==1.2.0
mpi4py==3.1.4
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
nibabel==5.1.0
ninja==1.11.1
nni==2.10.1
nptyping==2.5.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnx==1.14.0
onnxruntime==1.15.1
openslide-python==1.1.2
optuna==3.2.0
orderedmultidict==1.0.1
packaging==23.1
pandas==2.0.3
pathlib2==2.3.7.post1
pathspec==0.11.1
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.9.1
prettytable==3.8.0
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.18.0
pydicom==2.4.2
PyJWT==2.4.0
pynrrd==1.0.0
pynvml==11.5.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PythonWebHDFS==0.2.3
pytorch-fid==0.3.0
pytorch-ignite==0.4.11
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
querystring-parser==1.2.4
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.23.1
rpds-py==0.9.2
rsa==4.9
schema==0.7.5
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
SimpleITK==2.2.1
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.10
tokenizers==0.12.1
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers==4.21.3
triton==2.0.0
typeguard==2.13.3
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
zipp==3.16.2
